Refactored the structure and re-referenced the whole project.
Refactor project structure and update repository usage

Updated CreateStudentCommandHandler to use new interfaces from SchedsForums.Application. Modified SchedsForums.Application.csproj to reference SchedsForums.Domain. Simplified BaseEntity by removing IBaseEntity implementation. Refactored BaseUser, Admin, Moderator, and Student to remove base class constructor calls. Removed obsolete interfaces from SchedsForums.Domain.Interfaces. Reintroduced OnModelCreating in SchedsForumsDbContext. Replaced GenericRepository with BaseRepository. Updated StudentRepository to inherit from BaseRepository and use new IStudentRepository interface. Updated SchedsForums.Infrastructure.csproj to reference SchedsForums.Application. Updated HashingService to use IHashingService interface. Added new interfaces and BaseRepository class to SchedsForums.Application.Interfaces and SchedsForums.Infrastructure.Repositories.Common respectively.
