+++ SchedsForums.Application/Interfaces/Common/IBaseRepository.cs
@@ -1,0 +1,13 @@
+﻿using SchedsForums.Domain.Entities.Common;
+
+namespace SchedsForums.Application.Interfaces.Common
+{
+    public interface IBaseRepository<T> where T : BaseEntity
+    {
+        Task<T?> GetByIdAsync(string id);
+        Task<IEnumerable<T>> GetAllAsync();
+        Task<T> InsertAsync(T entity);
+        Task<T> UpdateAsync(T entity);
+        Task DeleteAsync(string id);
+    }
+}
