+++ SchedsForums.Application/Interfaces/Common/IBaseService.cs
@@ -1,0 +1,6 @@
+﻿namespace SchedsForums.Application.Interfaces.Common
+{
+    public interface IBaseService
+    {
+    }
+}
