+++ SchedsForums.Application/Interfaces/Repositories/IStudentRepository.cs
@@ -1,0 +1,9 @@
+﻿using SchedsForums.Application.Interfaces.Common;
+using SchedsForums.Domain.Entities.Users;
+
+namespace SchedsForums.Application.Interfaces.Repositories
+{
+    public interface IStudentRepository : IBaseRepository<Student>
+    {
+    }
+}
