+++ SchedsForums.Domain/Entities/Users/Admin.cs
@@ -4,11 +4,5 @@ namespace SchedsForums.Domain.Entities.Users
 {
     public class Admin : BaseUser
     {
-        public Admin(string name, string email, string password) : base(name, email, password)
-        {
-        }
-        protected Admin() : base()
-        {
-        }
     }
 }
