+++ SchedsForums.Domain/Entities/Users/Moderator.cs
@@ -4,11 +4,5 @@ namespace SchedsForums.Domain.Entities.Users
 {
     public class Moderator : BaseUser
     {
-        public Moderator(string name, string email, string password) : base(name, email, password)
-        {
-        }
-        public Moderator() : base()
-        {
-        }
     }
 }
