+++ SchedsForums.Domain/Entities/Users/Student.cs
@@ -4,9 +4,5 @@ namespace SchedsForums.Domain.Entities.Users
 {
     public class Student : BaseUser
     {
-        public Student(string name, string email, string password) : base(name, email, password)
-        {
-        }
-        public Student() : base() { }
     }
 }
