--- SchedsForums.Domain/Interfaces/IBaseEntity.cs
@@ -1,9 +1,0 @@
-﻿namespace SchedsForums.Domain.Interfaces
-{
-    public interface IBaseEntity
-    {
-        string Id { get; }
-        DateTime CreatedAt { get; }
-        DateTime UpdatedAt { get; }
-    }
-}
