--- SchedsForums.Domain/Interfaces/IBaseService.cs
@@ -1,6 +1,0 @@
-﻿namespace SchedsForums.Domain.Interfaces
-{
-    public interface IBaseService
-    {
-    }
-}
