--- SchedsForums.Domain/Interfaces/Repositories/IStudentRepository.cs
@@ -1,8 +1,0 @@
-﻿using SchedsForums.Domain.Entities.Users;
-
-namespace SchedsForums.Domain.Interfaces.Repositories
-{
-    public interface IStudentRepository : IGenericRepository<Student>
-    {
-    }
-}
