+++ SchedsForums.Infrastructure/Contexts/SchedsForumsDbContext.cs
@@ -5,14 +5,14 @@ namespace SchedsForums.Infrastructure.Contexts
 {
     public class SchedsForumsDbContext(DbContextOptions<SchedsForumsDbContext> options) : DbContext(options)
     {
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.ApplyConfigurationsFromAssembly(typeof(SchedsForumsDbContext).Assembly);
-            base.OnModelCreating(modelBuilder);
-        }
         public DbSet<BaseUser> Users { get; set; }
         public DbSet<Admin> Admins { get; set; }
         public DbSet<Student> Students { get; set; }
         public DbSet<Moderator> Moderators { get; set; }
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.ApplyConfigurationsFromAssembly(typeof(SchedsForumsDbContext).Assembly);
+            base.OnModelCreating(modelBuilder);
+        }   
     }
 }
