+++ SchedsForums.Infrastructure/Repositories/StudentRepository.cs
@@ -1,16 +1,16 @@
-﻿using SchedsForums.Domain.Entities.Users;
-using SchedsForums.Domain.Interfaces.Repositories;
+﻿using SchedsForums.Application.Interfaces.Repositories;
+using SchedsForums.Domain.Entities.Users;
 using SchedsForums.Infrastructure.Contexts;
 using SchedsForums.Infrastructure.Repositories.Common;
 
 namespace SchedsForums.Infrastructure.Repositories
 {
-    public class StudentRepository : GenericRepository<Student>, IStudentRepository
+    public class StudentRepository : BaseRepository<Student>, IStudentRepository
     {
         private readonly SchedsForumsDbContext _context;
         public StudentRepository(SchedsForumsDbContext context) : base(context)
         {
-            _context = context;
+            _context = context ?? throw new ArgumentNullException(nameof(context));
         }
     }
 }
