+++ SchedsForums.Infrastructure/SchedsForums.Infrastructure.csproj
@@ -20,6 +20,6 @@
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.8" />
     </ItemGroup>
     <ItemGroup>
-		<ProjectReference Include="..\SchedsForums.Domain\SchedsForums.Domain.csproj" />
+		<ProjectReference Include="..\SchedsForums.Application\SchedsForums.Application.csproj" />
 	</ItemGroup>
 </Project>
